Organize global search capture helpers
