+++ Sources/Search/GlobalSearchPanelCaptureManager.swift
@@ -669,135 +669,3 @@ final class GlobalSearchPanelCaptureManager {
         }
     }
 }
-
-/// Resolves each refresh waiter when one panel-content revision finishes or is superseded.
-@MainActor
-private final class GlobalSearchPanelCaptureCompletion {
-    let panelRevision: UInt64
-
-    private struct Waiter {
-        let continuation: CheckedContinuation<Bool, Never>
-        let deadline: GlobalSearchPanelCaptureDeadline
-    }
-
-    private var isFinished = false
-    private var waiters: [UUID: Waiter] = [:]
-
-    init(panelRevision: UInt64) {
-        self.panelRevision = panelRevision
-    }
-
-    func wait(until deadline: GlobalSearchPanelCaptureDeadline) async -> Bool {
-        guard !isFinished, !deadline.hasExpired, !Task.isCancelled else {
-            return isFinished && !Task.isCancelled
-        }
-        let waiterID = UUID()
-
-        return await withTaskCancellationHandler {
-            await withCheckedContinuation { continuation in
-                guard !isFinished, !deadline.hasExpired, !Task.isCancelled else {
-                    continuation.resume(returning: isFinished && !Task.isCancelled)
-                    return
-                }
-                waiters[waiterID] = Waiter(
-                    continuation: continuation,
-                    deadline: deadline
-                )
-                guard deadline.addExpirationHandler(
-                    id: waiterID,
-                    handler: { [weak self] in
-                        self?.expireWaiter(waiterID)
-                    }
-                ) else {
-                    waiters[waiterID] = nil
-                    continuation.resume(returning: false)
-                    return
-                }
-            }
-        } onCancel: {
-            Task { @MainActor [weak self] in
-                self?.cancelWaiter(waiterID)
-            }
-        }
-    }
-
-    func finish() {
-        guard !isFinished else { return }
-        isFinished = true
-        let pendingWaiters = waiters
-        waiters.removeAll()
-        for (waiterID, waiter) in pendingWaiters {
-            waiter.deadline.removeExpirationHandler(id: waiterID)
-            waiter.continuation.resume(returning: true)
-        }
-    }
-
-    private func cancelWaiter(_ waiterID: UUID) {
-        guard let waiter = waiters.removeValue(forKey: waiterID) else { return }
-        waiter.deadline.removeExpirationHandler(id: waiterID)
-        waiter.continuation.resume(returning: false)
-    }
-
-    private func expireWaiter(_ waiterID: UUID) {
-        guard let waiter = waiters.removeValue(forKey: waiterID) else { return }
-        waiter.continuation.resume(returning: false)
-    }
-}
-
-@MainActor
-private final class GlobalSearchPanelCaptureDeadline {
-    typealias ExpirationHandler = @MainActor () -> Void
-
-    private(set) var hasExpired = false
-    private var isCancelled = false
-    private var expirationHandlers: [UUID: ExpirationHandler] = [:]
-    private let timer: DispatchSourceTimer
-
-    init(milliseconds: Int) {
-        timer = DispatchSource.makeTimerSource(queue: .main)
-        timer.schedule(
-            deadline: .now() + .milliseconds(max(0, milliseconds)),
-            leeway: .milliseconds(10)
-        )
-        timer.setEventHandler { [weak self] in
-            MainActor.assumeIsolated {
-                self?.expire()
-            }
-        }
-        timer.resume()
-    }
-
-    func addExpirationHandler(
-        id: UUID,
-        handler: @escaping ExpirationHandler
-    ) -> Bool {
-        guard !hasExpired, !isCancelled else { return false }
-        expirationHandlers[id] = handler
-        return true
-    }
-
-    func removeExpirationHandler(id: UUID) {
-        expirationHandlers[id] = nil
-    }
-
-    func cancel() {
-        guard !isCancelled else { return }
-        isCancelled = true
-        expirationHandlers.removeAll()
-        timer.setEventHandler {}
-        timer.cancel()
-    }
-
-    private func expire() {
-        guard !hasExpired, !isCancelled else { return }
-        hasExpired = true
-        let handlers = Array(expirationHandlers.values)
-        expirationHandlers.removeAll()
-        timer.setEventHandler {}
-        timer.cancel()
-        isCancelled = true
-        for handler in handlers {
-            handler()
-        }
-    }
-}
