+++ Sources/Search/GlobalSearchPopoverSelectionUpdate.swift
@@ -1,0 +1,5 @@
+/// Describes how refreshed global-search results update the current selection.
+enum GlobalSearchPopoverSelectionUpdate: Equatable {
+    case reset
+    case preserveCurrentHit
+}
